[BUGFIX] Test with collapse field is broken with "64 bit numeric collapse fields are not supported" (#573)
This PR fixes the broken test for the variants against solr6

Reason: Only string and numeric fields should be used for field collapsing. In the testcase we used a text field.
This worked in Apache Solr 4.10.4 but not in 6.1.0 and was not intendet to work at all.

Since the error message in Apache Solr is missleading an Issue was created for that:

https://issues.apache.org/jira/browse/SOLR-9409

Because of that, the testcase was changed to use the numeric "pid" field for the grouping.